+++ app/controllers/students_controller.rb
@@ -4,23 +4,30 @@ class StudentsController < ApplicationController
   end
 
   def create
-    #raise
     @student = Student.new(stud_params)
 
     if @student.save
       redirect_to students_path
     else
-      render :new
+      render :new, :status => :bad_request
     end
   end
 
   def edit
-    @student = Student.find(params[:id])
+    @student = Student.find_by_id(params[:id])
+    if !@student
+      flash[:error] = "Student not found."
+      redirect_to students_path
+    end
   end
 
   def update
-    Student.update(params[:id], stud_params)
-    redirect_to students_path
+    @student = Student.update(params[:id], stud_params)
+    if @student.errors.any?
+      render :edit, :status => :bad_request
+    else
+      redirect_to students_path
+    end
   end
 
   def index
@@ -28,12 +35,21 @@ class StudentsController < ApplicationController
   end
 
   def show
-    @student = Student.find(params[:id])
+    @student = Student.find_by_id(params[:id])
+
+    if !@student
+      flash[:error] = "Student not found."
+      redirect_to students_path
+    end
   end
 
   def destroy
-    student = Student.find(params[:id])
-    student.destroy
+    student = Student.find_by_id(params[:id])
+    if student
+      student.destroy
+    else
+      flash[:error] = "Student not found."
+    end
     redirect_to students_path
   end
